split tests for FileSource to a separate file
